requirements.txt
Django>=4.2,<5.0
django-cors-headers
djangorestframework
psycopg2
black
drf-spectacular
Pillow
uwsgi